+++ src/modules/elastic_net/elastic_net_gaussian_fista.cpp
@@ -50,7 +50,7 @@ class GaussianFista
 inline void GaussianFista::update_y_intercept_final (
     FistaState<MutableArrayHandle<double> >& state)
 {
-    state.gradient_intercept = state.gradient_intercept / state.totalRows;
+    state.gradient_intercept = state.gradient_intercept / static_cast<double>(state.totalRows);
 }
 // -----------------------------------------------------------------------------
